+++ src/lib/audio/VolumeControl.svelte
@@ -8,28 +8,22 @@
 	// TODO use bind:value when signals adds `set`
 </script>
 
-<div class="volume-control">
-	<label>
-		<div class="title">volume</div>
-		<input
-			type="number"
-			on:input={(e) => (volume.value = Number(e.currentTarget.value))}
-			step={0.01}
-			min={0}
-			max={1}
-			value={$volume}
-		/>
-		<input
-			type="range"
-			on:input={(e) => (volume.value = Number(e.currentTarget.value))}
-			step={0.01}
-			min={0}
-			max={1}
-			value={$volume}
-		/>
-	</label>
-</div>
-
-<style>
-	/* .volume-control {} */
-</style>
+<label>
+	<div class="title">volume</div>
+	<input
+		type="number"
+		on:input={(e) => (volume.value = Number(e.currentTarget.value))}
+		step={0.01}
+		min={0}
+		max={1}
+		value={$volume}
+	/>
+	<input
+		type="range"
+		on:input={(e) => (volume.value = Number(e.currentTarget.value))}
+		step={0.01}
+		min={0}
+		max={1}
+		value={$volume}
+	/>
+</label>
