+++ src/lib/music/midi.ts
@@ -3,6 +3,14 @@ import type {Flavored} from '@feltjs/util';
 
 import type {Frequency} from '$lib/audio/helpers';
 import {DEFAULT_TUNING} from '$lib/music/helpers';
+import {
+	NOTE_SHARP_SYMBOL,
+	pitch_classes,
+	type Chroma,
+	type NoteName,
+	type Octave,
+	type PitchClass,
+} from '$lib/music/notes';
 
 // TODO min audible Midi (max too?)
 
@@ -39,3 +47,20 @@ export const freq_to_midi_safe = (freq: Frequency, tuning: Frequency): Midi | nu
 	if (!is_midi(midi)) return null;
 	return midi;
 };
+
+// TODO consider converting all of these to `Map`s
+// TODO do we want to remove the `midi` part of these data array names, or otherwise rename them?
+// maybe instead of `midiFoos`, rename to `noteFoos`?
+export const midi_chromas: Chroma[] = Object.freeze(midis.map((m) => (m % 12) + 1)) as Chroma[];
+export const midi_pitch_classes: PitchClass[] = Object.freeze(
+	midis.map((m) => pitch_classes[midi_chromas[m] - 1]),
+) as PitchClass[];
+export const midi_octaves: Octave[] = Object.freeze(
+	midis.map((m) => Math.floor(m / 12) - 1),
+) as Octave[];
+export const midi_names: NoteName[] = Object.freeze(
+	midis.map((m) => midi_pitch_classes[m] + midi_octaves[m]),
+) as NoteName[];
+export const midi_naturals: Set<Midi> = new Set(
+	midis.filter((m) => midi_pitch_classes[m][1] !== NOTE_SHARP_SYMBOL),
+);
