+++ src/lib/music/notes.ts
@@ -2,7 +2,7 @@ import {z} from 'zod';
 import type {Flavored} from '@feltjs/util';
 import {identity} from '@feltjs/util/function.js';
 
-import {type Midi, midis, is_midi} from '$lib/music/midi';
+import {type Midi, is_midi, midi_naturals} from '$lib/music/midi';
 import {type Hsl, hsl_to_string, type Hue} from '$lib/util/colors';
 
 // `Midi` is our primary means of identifying notes,
@@ -58,23 +58,6 @@ export const parse_intervals = (value: string): Intervals =>
 		.split(',')
 		.map((v) => Number(v.trim()) | 0)
 		.filter(Boolean); // exclude 0 intentionally
-
-// TODO consider converting all of these to `Map`s
-// TODO do we want to remove the `midi` part of these data array names, or otherwise rename them?
-// maybe instead of `midiFoos`, rename to `noteFoos`?
-export const midi_chromas: Chroma[] = Object.freeze(midis.map((m) => (m % 12) + 1)) as Chroma[];
-export const midi_pitch_classes: PitchClass[] = Object.freeze(
-	midis.map((m) => pitch_classes[midi_chromas[m] - 1]),
-) as PitchClass[];
-export const midi_octaves: Octave[] = Object.freeze(
-	midis.map((m) => Math.floor(m / 12) - 1),
-) as Octave[];
-export const midi_names: NoteName[] = Object.freeze(
-	midis.map((m) => midi_pitch_classes[m] + midi_octaves[m]),
-) as NoteName[];
-export const midi_naturals: Set<Midi> = new Set(
-	midis.filter((m) => midi_pitch_classes[m][1] !== NOTE_SHARP_SYMBOL),
-);
 
 export const compute_naturals = (min: Midi, max: Midi): Midi[] => {
 	const naturals: Midi[] = [];
