+++ src/lib/music/piano.ts
@@ -1,5 +1,6 @@
 import type {Midi} from '$lib/music/midi';
-import {compute_naturals, midi_chromas, midi_naturals, type Chroma} from '$lib/music/notes';
+import {midi_chromas, midi_naturals} from '$lib/music/midi';
+import {compute_naturals, type Chroma} from '$lib/music/notes';
 
 const KEY_HEIGHT_MULT = 5; // width * mult = height // TODO - make dependent on container?
 const ACCIDENTAL_KEY_WIDTH_MULT = 7 / 12;
